Refactor team form
